feat(examples): 添加虚拟挖矿器和 F2Pool 示例配置
- 新增多个虚拟挖矿器示例，包括单核、多核和 F2Pool 配置
- 添加 F2Pool 简化配置和虚拟挖矿配置
- 实现基本的虚拟挖矿功能，用于测试和演示
- 添加认证中间件实现

+++ src/device/chain.rs
@@ -365,7 +365,7 @@ impl ChainController for AsicChainController {
         }
 
         info!("Detected {} chips on chain {}", chip_count, self.chain_id);
-        Ok(chip_count)
+        Ok(chip_count as u32)
     }
 
     /// 设置PLL频率
